saving progress on dev branch

+++ src/Repository/UserRepository.php
@@ -12,6 +12,7 @@ use Symfony\Component\Security\Core\User\PasswordUpgraderInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Form\Extension\Core\Type\DateTimeType;
 use DateTime;
+use App\Repository\AccountsDurationRepository;
 
 /**
  * @method User|null find($id, $lockMode = null, $lockVersion = null)
@@ -21,9 +22,15 @@ use DateTime;
  */
 class UserRepository extends ServiceEntityRepository implements PasswordUpgraderInterface
 {
-    public function __construct(ManagerRegistry $registry)
+    /**
+     * @var \App\Repository\AccountsDurationRepository
+     */
+    private $aDuRepo;
+
+    public function __construct(ManagerRegistry $registry, AccountsDurationRepository $aDuRepo)
     {
         parent::__construct($registry, User::class);
+        $this->aDuRepo = $aDuRepo;
     }
 
     /**
@@ -68,15 +75,19 @@ class UserRepository extends ServiceEntityRepository implements PasswordUpgrader
      */
     public function deleteOldAccounts(): void
     {
-        $now = new DateTime();
-        $before = date_modify($now, '-100 days');
+        $accountsDuration = $this->aDuRepo->findOneBy([]);
+        if ($accountsDuration) {
+            $days = $accountsDuration->getDays();
+            $now = new DateTime();
+            $before = date_modify($now, "-$days days");
 
-        $qb = $this->getEntityManager()->createQueryBuilder();
+            $qb = $this->getEntityManager()->createQueryBuilder();
 
-        $qb->delete(User::class, 'u')
-            ->where("u.createdAt < :before")
-            ->setParameter('before', $before, \Doctrine\DBAL\Types\Types::DATETIME_MUTABLE)
-            ->getQuery()->execute();
+            $qb->delete(User::class, 'u')
+                ->where("u.createdAt < :before")
+                ->setParameter('before', $before, \Doctrine\DBAL\Types\Types::DATETIME_MUTABLE)
+                ->getQuery()->execute();
+        }
     }
 
     // /**
